--- src/main/java/ddingdong/ddingdongBE/domain/form/service/FacadeUserFormService.java
@@ -1,9 +1,0 @@
-package ddingdong.ddingdongBE.domain.form.service;
-
-import ddingdong.ddingdongBE.domain.form.service.dto.command.CreateFormResponseCommand;
-
-public interface FacadeUserFormService {
-
-    void createFormResponse(CreateFormResponseCommand createFormResponseCommand);
-
-}
