--- src/main/java/ddingdong/ddingdongBE/domain/form/service/FormAnswerService.java
@@ -1,10 +1,0 @@
-package ddingdong.ddingdongBE.domain.form.service;
-
-import ddingdong.ddingdongBE.domain.form.entity.FormAnswer;
-import java.util.List;
-
-public interface FormAnswerService {
-
-    void createAll(List<FormAnswer> formAnswers);
-
-}
