--- src/main/java/ddingdong/ddingdongBE/domain/form/service/FormResponseService.java
@@ -1,9 +1,0 @@
-package ddingdong.ddingdongBE.domain.form.service;
-
-import ddingdong.ddingdongBE.domain.form.entity.FormResponse;
-
-public interface FormResponseService {
-
-    FormResponse create(FormResponse formResponse);
-
-}
