+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/controller/UserFormController.java
@@ -1,7 +1,7 @@
-package ddingdong.ddingdongBE.domain.form.controller.dto;
+package ddingdong.ddingdongBE.domain.formapplicaion.controller;
 
-import ddingdong.ddingdongBE.domain.form.api.UserFormApi;
-import ddingdong.ddingdongBE.domain.form.controller.dto.request.CreateFormResponseRequest;
+import ddingdong.ddingdongBE.domain.formapplicaion.api.UserFormApi;
+import ddingdong.ddingdongBE.domain.formapplicaion.controller.dto.request.CreateFormResponseRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
