+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/entity/FormResponse.java
@@ -1,13 +1,12 @@
-package ddingdong.ddingdongBE.domain.form.entity;
+package ddingdong.ddingdongBE.domain.formapplicaion.entity;
 
 import ddingdong.ddingdongBE.common.BaseEntity;
+import ddingdong.ddingdongBE.domain.form.entity.Form;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -17,9 +16,6 @@ public class FormResponse extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(nullable = false)
-    private LocalDateTime submittedAt;
 
     @Column(nullable = false)
     private String name;
@@ -38,8 +34,7 @@ public class FormResponse extends BaseEntity {
     private Form form;
 
     @Builder
-    private FormResponse(LocalDateTime submittedAt, String name, String studentNumber, String department, FormResponseStatus status, Form form) {
-        this.submittedAt = submittedAt;
+    private FormResponse(String name, String studentNumber, String department, FormResponseStatus status, Form form) {
         this.name = name;
         this.studentNumber = studentNumber;
         this.department = department;
