+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/entity/FormResponseStatus.java
@@ -1,4 +1,4 @@
-package ddingdong.ddingdongBE.domain.form.entity;
+package ddingdong.ddingdongBE.domain.formapplicaion.entity;
 
 public enum FormResponseStatus {
     SUBMITTED,
