+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/repository/FormResponseRepository.java
@@ -1,0 +1,7 @@
+package ddingdong.ddingdongBE.domain.formapplicaion.repository;
+
+import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormResponse;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FormResponseRepository extends JpaRepository<FormResponse, Long> {
+}
