+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/service/FormAnswerService.java
@@ -1,0 +1,10 @@
+package ddingdong.ddingdongBE.domain.formapplicaion.service;
+
+import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormAnswer;
+import java.util.List;
+
+public interface FormAnswerService {
+
+    void createAll(List<FormAnswer> formAnswers);
+
+}
