+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/service/FormResponseService.java
@@ -1,0 +1,9 @@
+package ddingdong.ddingdongBE.domain.formapplicaion.service;
+
+import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormResponse;
+
+public interface FormResponseService {
+
+    FormResponse create(FormResponse formResponse);
+
+}
